pyproject.toml
[project]
name = "fortune-telling-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "lunarcalendar>=0.0.9",
    "lunar_python>=1.4.0",
    "jinja2>=3.1.6",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "isort>=8.0.1",
    "mypy>=1.19.1",
    "pylint>=4.0.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
]

[tool.ruff]
extend-exclude = [
    "venv",
    "__pycache__",
    ".git",
    ".venv",
    "meishiki",
    "numerology",
]
lint.ignore = ["E402"]

[tool.pylint]
ignore-paths = [
    "venv/",
    "__pycache__/",
    ".git/",
    ".venv",
    "meishiki",
    "numerology",
]

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "C0413",  # wrong-import-order
    "E0401",  # import-error
]

[tool.mypy]
exclude = [
    "venv/",
    "__pycache__/",
    ".git/",
    ".venv",
    "meishiki",
    "numerology",
]
ignore_missing_imports = true
disable_error_code = ["attr-defined"]

[tool.pytest.ini_options]
pythonpath = [".", "./meishiki", "./numerology"]
norecursedirs = "numerology"
